control/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
deprecation==2.1.0
Flask==3.1.0
functions-framework==3.8.2
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.0
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
